Comply with ClassLength
